% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_dt.R
\name{fct_to_int}
\alias{fct_to_int}
\title{Convert Factor to Integer}
\usage{
fct_to_int(x, byref = FALSE)
}
\arguments{
\item{x}{A factor variable to be converted to an integer vector.}

\item{byref}{Logical. If TRUE, the conversion is done in-place; if FALSE, a new object is created.}
}
\value{
An integer vector obtained by converting the factor variable to integers.
}
\description{
This function converts a factor variable to an integer vector. If the \code{byref} parameter
is set to \code{FALSE} (default), a new object is created; otherwise, the conversion is done in-place.
}
