% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{generate_corr_unifs}
\alias{generate_corr_unifs}
\title{Given a correlation matrix (Pearson), produces a matrix of correlated uniforms}
\usage{
generate_corr_unifs(n, M)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{M}{The correlation matrix.}
}
\value{
A matrix of correlated uniform random variables.
}
\description{
This function generates correlated uniform random variables based on a given correlation matrix using the dqrnorm and pnorm functions.
}
