% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{generate_rns}
\alias{generate_rns}
\title{Generate Random Numbers for Monte Carlo iteration}
\usage{
generate_rns(mc, dt, colnams)
}
\arguments{
\item{mc}{A numeric value representing the Monte Carlo iteration run.}

\item{dt}{A data table containing the columns to be updated with random numbers.}

\item{colnams}{A character vector specifying the names of columns to be updated with random numbers.}
}
\value{
The input data table with specified columns updated with random numbers.
}
\description{
This function generates random numbers for a Monte Carlo iteration using the dqRNG package. It sets the seed based on the simulation run and updates specified columns in the input data table with random numbers.
}
