% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghAssetUtils.R
\name{GetGitHubAssetRouteInfo}
\alias{GetGitHubAssetRouteInfo}
\title{Find GitHub assets route information}
\usage{
GetGitHubAssetRouteInfo(
  sRepo,
  sTag,
  sUploadSrcDirPath,
  sDeployToRootDirPath,
  bOverwriteFilesOnDeploy,
  sToken = NULL,
  iTestWithFirstNAssets
)
}
\arguments{
\item{sRepo}{string (out param) GitHub repository name.}

\item{sTag}{string (out param) GitHub repository tag.}

\item{sUploadSrcDirPath}{string (out param): source directory path to scan for uploading assets to GitHub.}

\item{sDeployToRootDirPath}{string (out param): deployment directory path for downloading assets from GitHub.}

\item{bOverwriteFilesOnDeploy}{bool (out param): overwrite files during deployment.}

\item{sToken}{string (in|out param): GitHub personal access token (PAT).}

\item{iTestWithFirstNAssets}{int (out param): only download first m assets (for testing).}
}
\description{
Get GitHub asset route data from command-line variables.
}
