% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghAssetUtils.R
\name{HttpHeaderNamesAndValues}
\alias{HttpHeaderNamesAndValues}
\title{HTTP header names and values}
\usage{
HttpHeaderNamesAndValues(httpResponse)
}
\arguments{
\item{httpResponse}{HTTP response object as provided by \link{httr} package.}
}
\value{
String giving each header's name and value.
}
\description{
Helper to create descriptive list of HTTP headers from the given response object.
}
