% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACTncdEngl.R
\docType{package}
\name{IMPACTncdEngl}
\alias{IMPACTncdEngl}
\title{IMPACTncdEngl: A package with functions for the IMPACTncdEngl model.}
\description{
Description of your package.
}
\author{
Chris Kypridemos
}
