% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACTncdEnglmisc.R
\docType{package}
\name{IMPACTncdEnglmisc}
\alias{IMPACTncdEnglmisc}
\title{IMPACTncdEnglmisc: A package with functions for the IMPACTncdEngl model.}
\description{
Description of your package.
}
\author{
Chris Kypridemos
}
