% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_dt.R
\name{is_valid_lookup_tbl}
\alias{is_valid_lookup_tbl}
\title{Check Validity of Lookup Table}
\usage{
is_valid_lookup_tbl(lookup_tbl, keycols)
}
\arguments{
\item{lookup_tbl}{The data.table representing the lookup table.}

\item{keycols}{A character vector specifying the key columns in the lookup table.}
}
\value{
TRUE if the lookup table is valid; otherwise, an error is raised.
}
\description{
This function checks the validity of a lookup table, ensuring that it meets the required conditions
for key columns. It verifies that key columns are unique, of type integer or factor, and contain
the necessary integer values between min and max for integer keys.
}
