% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{mean_predictAll}
\alias{mean_predictAll}
\title{Predict Mean Values from a GAMLSS Model for Multiple Observations}
\usage{
mean_predictAll(gamlss_model, orig_data, newdata)
}
\arguments{
\item{gamlss_model}{A GAMLSS model object created using the \code{gamlss} package.}

\item{orig_data}{The original data used to fit the GAMLSS model.}

\item{newdata}{The new data for which predictions are to be made.}
}
\value{
A numeric vector containing the predicted mean values for each observation in \code{newdata}.
}
\description{
This function predicts mean values from a Generalized Additive Model for Location, Scale, and Shape (GAMLSS) model for multiple observations.
}
