% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_class.R
\name{Parameter}
\alias{Parameter}
\title{R6 Class representing a parameter}
\description{
A parameter object has a stochastic simulation parameter.
}
\details{
To be completed...
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the parameter.}

\item{\code{outcome}}{The name of the outcome the parameter influences.}

\item{\code{distribution}}{The distribution to be used for Monte Carlo.
Currently lognormal and normal are supported.}

\item{\code{source}}{Citation info for the effect size.}

\item{\code{notes}}{Any notes regarding the parameter -> outcome relation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Parameter$new()}}
\item \href{#method-read_exps_prm}{\code{Parameter$read_exps_prm()}}
\item \href{#method-gen_stochastic_effect}{\code{Parameter$gen_stochastic_effect()}}
\item \href{#method-get_rr}{\code{Parameter$get_rr()}}
\item \href{#method-clear_cache}{\code{Parameter$clear_cache()}}
\item \href{#method-validate_rr}{\code{Parameter$validate_rr()}}
\item \href{#method-get_input_rr}{\code{Parameter$get_input_rr()}}
\item \href{#method-get_metadata}{\code{Parameter$get_metadata()}}
\item \href{#method-get_seed}{\code{Parameter$get_seed()}}
\item \href{#method-write_xps_tmplte_file}{\code{Parameter$write_xps_tmplte_file()}}
\item \href{#method-convert_from_old_format}{\code{Parameter$convert_from_old_format()}}
\item \href{#method-print}{\code{Parameter$print()}}
\item \href{#method-clone}{\code{Parameter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new parameter object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$new(
  exps_name,
  outcome,
  distribution = c("lognormal", "normal"),
  source = NA_character_,
  notes = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exps_name}}{A string with parameter name.}

\item{\code{outcome}}{A string with outcome name.}

\item{\code{distribution}}{A string with distribution for Monte Carlo (lognormal or normal).}

\item{\code{source}}{A string with source info.}

\item{\code{notes}}{A string with any notes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Parameter` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_exps_prm"></a>}}
\if{latex}{\out{\hypertarget{method-read_exps_prm}{}}}
\subsection{Method \code{read_exps_prm()}}{
Reads parameter parameter from file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$read_exps_prm(exps_prm, design_ = design)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exps_prm}}{A path to a csvy file with the parameter parameters.}

\item{\code{design_}}{A design object with the simulation parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `Parameter` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_stochastic_effect"></a>}}
\if{latex}{\out{\hypertarget{method-gen_stochastic_effect}{}}}
\subsection{Method \code{gen_stochastic_effect()}}{
Generates and write to disk the stochastic effect.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$gen_stochastic_effect(
  design_ = design,
  overwrite = FALSE,
  smooth,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{design_}}{A design object with the simulation parameters.}

\item{\code{overwrite}}{If TRUE overwrite the files. Else if files exist they
are not regenerated.}

\item{\code{smooth}}{If true applies loess smoothing to the input relative risks}

\item{\code{...}}{Further arguments to be passed to `loess`, usually span =
0.7, degree = 1}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `Parameter` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rr"></a>}}
\if{latex}{\out{\hypertarget{method-get_rr}{}}}
\subsection{Method \code{get_rr()}}{
Get relative risks from disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$get_rr(mc, design_ = design, drop = TRUE, plot_rr = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mc}}{An integer that signifies the Monte Carlo iteration.}

\item{\code{design_}}{A design object with the simulation parameters.}

\item{\code{drop}}{If `TRUE` returns a scalar numeric of the RR if the RR is
common for all age groups and both sexes. Otherwise, a data.table.}

\item{\code{plot_rr}}{If TRUE, plots the relative risk}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table with the stochastic relative risks, if stochastic
  = TRUE; else, the deterministic relative risks.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_cache"></a>}}
\if{latex}{\out{\hypertarget{method-clear_cache}{}}}
\subsection{Method \code{clear_cache()}}{
Clear the cache for get_rr.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$clear_cache()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The `Parameter` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_rr"></a>}}
\if{latex}{\out{\hypertarget{method-validate_rr}{}}}
\subsection{Method \code{validate_rr()}}{
Get relative risks from disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$validate_rr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A plot with the input and stochastic relative risks.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_input_rr"></a>}}
\if{latex}{\out{\hypertarget{method-get_input_rr}{}}}
\subsection{Method \code{get_input_rr()}}{
Get original input relative risks by age and sex.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$get_input_rr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A copied data.table with the original relative risks.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Get original metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$get_metadata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list with the original metadata.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_seed"></a>}}
\if{latex}{\out{\hypertarget{method-get_seed}{}}}
\subsection{Method \code{get_seed()}}{
Get seed for RNG.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$get_seed()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A seed for the RNG that is produced by the digest of parameter
  name and outcome.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_xps_tmplte_file"></a>}}
\if{latex}{\out{\hypertarget{method-write_xps_tmplte_file}{}}}
\subsection{Method \code{write_xps_tmplte_file()}}{
Writes a template parameter file to disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$write_xps_tmplte_file(file_path = "./inputs/RR/template.csvy")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_path}}{Path including file name and .csvy extension to write the
file with placeholder parameter parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The `Parameter` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convert_from_old_format"></a>}}
\if{latex}{\out{\hypertarget{method-convert_from_old_format}{}}}
\subsection{Method \code{convert_from_old_format()}}{
Convert the old format .csv to the new format .csvy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$convert_from_old_format(
  old_file,
  metadata,
  file_path,
  estimates = NULL,
  second_estimate_is_p = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{old_file}}{Path to the old format .csv file with the RR.}

\item{\code{metadata}}{List with the metadata information.}

\item{\code{file_path}}{Path including file name and .csvy extension to write the
new file.}

\item{\code{estimates}}{Only used when old_file is missing. A vector of length
2 with the 1st element being the point estimate and the 2nd one of
the CI.}

\item{\code{second_estimate_is_p}}{If `TRUE` and the `estimates` not `NULL` then the
second element of `estimates` is interpreted like a p-value for ratio
rather than a CI.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The `Parameter` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the simulation parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The `Parameter` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
