% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghAssetUtils.R
\name{StopOnHttpFailure}
\alias{StopOnHttpFailure}
\title{Stop on failure}
\usage{
StopOnHttpFailure(lsHttpResponses, bUploadedFiles)
}
\arguments{
\item{lsHttpResponses}{List of HTTP response objects, each provided by httr package.}

\item{bUploadedFiles}{boolean, files have been uploaded.}
}
\description{
Stop if get HTTP response indicating missing file or unexpected file size.
}
