% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghAssetUtils.R
\name{TrimSlashes}
\alias{TrimSlashes}
\title{Trim slashes}
\usage{
TrimSlashes(sPath, bRidStartSlash = TRUE)
}
\arguments{
\item{sPath:}{string initial path.}

\item{bRidStartSlash:}{bool remove slash at start of path}
}
\value{
string possibly modified path.
}
\description{
Remove slashes from start and end of path.
}
