% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACTncdEngl.R
\docType{package}
\name{IMPACTncdEngl}
\alias{-package}
\alias{IMPACTncdEngl}
\alias{.datatable.aware}
\title{IMPACTncdEngl: A package with functions for the IMPACTncdEngl model.}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\description{
Description of your package.
}
\author{
Chris Kypridemos
}
\keyword{datasets}
